requirements.txt
ccxt==4.1.77
python-binance==1.0.17
pandas==2.2.1
numpy==1.26.4
ta==0.11.0
python-telegram-bot==20.8
requests==2.31.0
python-dotenv==1.0.1
schedule==1.2.1
matplotlib>=3.7.0